Finished version of the game. Has everything in it.
Can be modified to be a bit more organized, but all javadoc is there so it is pretty much finished. There are some methods that are used for testing that we could remove, but I probably want to use them to show Gilmore so don't delete if you modify this version. If you make any modifications, please don't have them change the code, everything right now is decently tested and small edits to code could break it without us noticing.